add branchPrefix "deps-" to make docker tags valid
when CI/CD deployment pipelines run they use the branch name to push docker tags. They must be a valid format for docker tags.
By default renovate creates branches with name "renovate/<libname><version>" which will not be a valid format. 
With prefix deps- the tag will be "deps-lib-version" instead which is valid for docker tags.